+++ test/functional_cpu_test.py
@@ -7,35 +7,17 @@ import torchaudio.functional as F
 import pytest
 
 from . import common_utils
+from .functional_impl import Lfilter
 
 
-class Lfilter(common_utils.TestBaseMixin):
-    def test_simple(self):
-        """
-        Create a very basic signal,
-        Then make a simple 4th order delay
-        The output should be same as the input but shifted
-        """
-
-        torch.random.manual_seed(42)
-        waveform = torch.rand(2, 44100 * 1, dtype=self.dtype, device=self.device)
-        b_coeffs = torch.tensor([0, 0, 0, 1], dtype=self.dtype, device=self.device)
-        a_coeffs = torch.tensor([1, 0, 0, 0], dtype=self.dtype, device=self.device)
-        output_waveform = F.lfilter(waveform, a_coeffs, b_coeffs)
-
-        self.assertEqual(output_waveform[:, 3:], waveform[:, 0:-3], atol=1e-5, rtol=1e-5)
-
-    def test_clamp(self):
-        input_signal = torch.ones(1, 44100 * 1, dtype=self.dtype, device=self.device)
-        b_coeffs = torch.tensor([1, 0], dtype=self.dtype, device=self.device)
-        a_coeffs = torch.tensor([1, -0.95], dtype=self.dtype, device=self.device)
-        output_signal = F.lfilter(input_signal, a_coeffs, b_coeffs, clamp=True)
-        assert output_signal.max() <= 1
-        output_signal = F.lfilter(input_signal, a_coeffs, b_coeffs, clamp=False)
-        assert output_signal.max() > 1
+class TestLFilterFloat32(Lfilter, common_utils.TestCase):
+    dtype = torch.float32
+    device = torch.device('cpu')
 
 
-common_utils.define_test_suites(globals(), [Lfilter])
+class TestLFilterFloat64(Lfilter, common_utils.TestCase):
+    dtype = torch.float64
+    device = torch.device('cpu')
 
 
 class TestComputeDeltas(unittest.TestCase):
