+++ test/functional_cuda_test.py
@@ -1,16 +1,16 @@
 import torch
 
 from . import common_utils
-from .kaldi_compatibility_impl import Kaldi
+from .functional_impl import Lfilter
 
 
 @common_utils.skipIfNoCuda
-class TestKaldiFloat32(Kaldi, common_utils.TestCase):
+class TestLFilterFloat32(Lfilter, common_utils.TestCase):
     dtype = torch.float32
     device = torch.device('cuda')
 
 
 @common_utils.skipIfNoCuda
-class TestKaldiFloat64(Kaldi, common_utils.TestCase):
+class TestLFilterFloat64(Lfilter, common_utils.TestCase):
     dtype = torch.float64
     device = torch.device('cuda')
