+++ test/kaldi_compatibility_cpu_test.py
@@ -1,5 +1,14 @@
+import torch
+
 from . import common_utils
 from .kaldi_compatibility_impl import Kaldi
 
 
-common_utils.define_test_suites(globals(), [Kaldi], devices=['cpu'])
+class TestKaldiFloat32(Kaldi, common_utils.TestCase):
+    dtype = torch.float32
+    device = torch.device('cpu')
+
+
+class TestKaldiFloat64(Kaldi, common_utils.TestCase):
+    dtype = torch.float64
+    device = torch.device('cpu')
