+++ test/torchscript_consistency_cpu_test.py
@@ -1,17 +1,24 @@
-from parameterized import parameterized_class
+import torch
 
-from .common_utils import TestCase, common_test_class_parameters
+from . import common_utils
 from .torchscript_consistency_impl import Functional, Transforms
 
 
-parameters = list(common_test_class_parameters(devices=['cpu']))
+class TestFunctionalFloat32(Functional, common_utils.TestCase):
+    dtype = torch.float32
+    device = torch.device('cpu')
 
 
-@parameterized_class(parameters)
-class TestFunctional(Functional, TestCase):
-    pass
+class TestFunctionalFloat64(Functional, common_utils.TestCase):
+    dtype = torch.float64
+    device = torch.device('cpu')
 
 
-@parameterized_class(parameters)
-class TestTransforms(Transforms, TestCase):
-    pass
+class TestTransformsFloat32(Transforms, common_utils.TestCase):
+    dtype = torch.float32
+    device = torch.device('cpu')
+
+
+class TestTransformsFloat64(Transforms, common_utils.TestCase):
+    dtype = torch.float64
+    device = torch.device('cpu')
